[wsgi] Allow configuration via Apache SetEnv directives
When running via mod_wsgi, there is no (easy) way to pass
configuration information at the point of application startup.  It is
possible to use SetEnv directives to specify environment variables,
but these are made available only as part of the first WSGI request.

Refactor the code to defer the filesystem scan until the first
request, thereby allowing the root directory to be specified via
SetEnv.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>